containers-data index

Library containers-data

This library exposes the following toplevel modules:

CCBV Imperative Bitvectors.

CCBijection Functor to build a bijection Represents 1-to-1 mappings between two types. Each element from the "left" is mapped to one "right" value, and conversely.

CCBitField Efficient Bit Field for up to 31 or 61 fiels

CCCache Caches Utils

CCDeque Imperative deque

CCFQueue Functional queues

CCFun_vec Functional Vectors

CCGraph Simple Graph Interface

CCHashSet Mutable Set

CCHashTrie Hash Tries

CCHet Associative containers with Heterogeneous Values

CCImmutArray Immutable Arrays

CCIntMap Map specialized for Int keys

CCKTree Lazy Tree Structure This structure can be used to represent trees and directed graphs (as infinite trees) in a lazy fashion. It is a structural type.

CCLazy_list Lazy List

CCMixmap Maps with Heterogeneous Values

CCMixset Set of Heterogeneous Values

CCMixtbl Hash Table with Heterogeneous Keys

CCMultiMap Map that can map key to several values

CCMultiSet Multiset

CCMutHeap Mutable Heaps

CCMutHeap_intf

CCPersistentArray Persistent Arrays

CCPersistentHashtbl Persistent hash-table on top of OCaml's hashtables

CCRAL Random-Access Lists

CCRingBuffer Circular Buffer (Deque)

CCSimple_queue Functional queues (fifo)

CCTrie Prefix Tree

CCWBTree Weight-Balanced Tree

CCZipper List Zipper

Library containers-data.top

The entry point of this library is the module: Containers_data_top.