containers index

Library containers

This library exposes the following toplevel modules:

CCArray Array utils

CCArrayLabels Array utils (Labeled version of CCArray)

CCAtomic

CCBool Basic Bool functions

CCByte_buffer Byte buffer.

CCCanonical_sexp Canonical S-expressions

CCChar Utils around char

CCEither Either Monad

CCEqual Equality Combinators

CCEqualLabels Equality Combinators (Labeled version of CCEqual)

CCFloat Basic operations on floating-point numbers

CCFormat Helpers for Format

CCFun Basic operations on Functions

CCHash Hash combinators

CCHashtbl Extension to the standard Hashtbl

CCHeap Leftist Heaps

CCIO IO Utils

CCInt Basic Int functions

CCInt32 Helpers for 32-bit integers.

CCInt64 Helpers for 64-bit integers.

CCList Complements to List

CCListLabels Complements to ListLabels

CCMap Extensions of Standard Map

CCNativeint Helpers for processor-native integers

CCOpt Previous Option module

CCOption Basic operations on the option type.

CCOrd Order combinators

CCPair Tuple Functions

CCParse Very Simple Parser Combinators

CCRandom Random Generators

CCRef Helpers for references

CCResult Error Monad

CCSeq Helpers for the standard Seq type

CCSet Wrapper around Set

CCSexp Handling S-expressions

CCSexp_intf

CCSexp_lex

CCShims_

CCShims_syntax

CCString Basic String Utils

CCStringLabels Basic String Utils (Labeled version of CCString)

CCUnit

CCUtf8_string Unicode String, in UTF8

CCVector Growable, mutable vector

Containers Drop-In replacement to Stdlib

ContainersLabels Drop-In replacement to Stdlib

Library containers.bencode

The entry point of this library is the module: Containers_bencode.

Library containers.cbor

The entry point of this library is the module: Containers_cbor.

Library containers.codegen

The entry point of this library is the module: Containers_codegen.

Library containers.monomorphic

The entry point of this library is the module: CCMonomorphic.

Library containers.scc

The entry point of this library is the module: Containers_scc.

Library containers.top

The entry point of this library is the module: Containers_top.

Library containers.unix

The entry point of this library is the module: CCUnix.